+++ lib/ex_gram/model/subtype/input_message_content.ex
@@ -1,6 +1,7 @@
 defimpl ExGram.Model.Subtype, for: ExGram.Model.InputMessageContent do
   def selector_value(_, params) do
     cond do
+      params[:rich_message] -> "rich"
       params[:message_text] -> "text"
       params[:address] -> "venue"
       params[:latitude] -> "location"
@@ -10,6 +11,7 @@ defimpl ExGram.Model.Subtype, for: ExGram.Model.InputMessageContent do
   end
 
   def subtype(_, "text"), do: ExGram.Model.InputTextMessageContent
+  def subtype(_, "rich"), do: ExGram.Model.InputRichMessageContent
   def subtype(_, "location"), do: ExGram.Model.InputLocationMessageContent
   def subtype(_, "venue"), do: ExGram.Model.InputVenueMessageContent
   def subtype(_, "contact"), do: ExGram.Model.InputContactMessageContent